Fix mishandling of arrays in MergeSort's preComputeManualSort() method
A new field "inputArrayDuplicate" has been added, which is a duplicate of inputArray, and mergeSortForManualSorter() operates on this array. This mirrors the behaviour of other preComputeManualSort() methods, which leave the original array untouched.

ManualSorter was being passed an already-sorted array due to a missing array copy during the initialisation of the ManualSorter object. This has been added and now this functions as intended.